+++ example/http/server/stackless-ssl/http_server_stackless_ssl.cpp
@@ -238,7 +238,7 @@ fail(beast::error_code ec, char const* what)
 
 // Handles an HTTP server connection
 class session
-    : public net::coroutine
+    : public boost::asio::coroutine
     , public std::enable_shared_from_this<session>
 {
     // This is the C++11 equivalent of a generic lambda.
@@ -398,7 +398,7 @@ public:
 
 // Accepts incoming connections and launches the sessions
 class listener
-    : public net::coroutine
+    : public boost::asio::coroutine
     , public std::enable_shared_from_this<listener>
 {
     net::io_context& ioc_;