+++ example/websocket/server/stackless-ssl/websocket_server_stackless_ssl.cpp
@@ -48,7 +48,7 @@ fail(beast::error_code ec, char const* what)
 
 // Echoes back all received WebSocket messages
 class session
-    : public net::coroutine
+    : public boost::asio::coroutine
     , public std::enable_shared_from_this<session>
 {
     websocket::stream<beast::ssl_stream<beast::tcp_stream>> ws_;
@@ -164,7 +164,7 @@ public:
 
 // Accepts incoming connections and launches the sessions
 class listener
-    : public net::coroutine
+    : public boost::asio::coroutine
     , public std::enable_shared_from_this<listener>
 {
     net::io_context& ioc_;